backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-multipart
Pillow
opencv-python-headless
python-dotenv
requests
aiofiles
pydantic[email]
numpy
passlib[bcrypt]
python-jose[cryptography]
